+++ strawtubes/strawtubesHit.cxx
@@ -50,7 +50,6 @@ void strawtubesHit::StrawEndPoints(TVector3 &vbot, TVector3 &vtop)
     Int_t pnb =  (fDetectorID- statnb*10000000 - vnb*1000000)/100000;
     Int_t lnb =  (fDetectorID - statnb*10000000 - vnb*1000000 - pnb*100000)/10000;
     TString stat = "Tr";stat+=+statnb;stat+="_";stat+=statnb;
-    if (statnb==5){stat="Veto_5";}
     TString view;
     switch (vnb) {
 	      case 0:
@@ -70,13 +69,11 @@ void strawtubesHit::StrawEndPoints(TVector3 &vbot, TVector3 &vtop)
 	        view = "_x1";}
     TGeoNavigator* nav = gGeoManager->GetCurrentNavigator();
     TString prefix = "Tr";
-    if (statnb==5){prefix="Veto";}
-    else{prefix+=statnb;}
+    prefix+=statnb;
     prefix+=view;prefix+="_plane_";prefix+=pnb;prefix+="_";
     TString plane = prefix;plane+=statnb;plane+=vnb;plane+=+pnb;plane+="00000";
     TString layer = prefix+"layer_";layer+=lnb;layer+="_";layer+=statnb;layer+=vnb;layer+=pnb;layer+=lnb;layer+="0000";
     TString wire = "wire_";
-    if (statnb==5){wire+="veto_";}
     wire+=(fDetectorID+1000);
     if (statnb<3){wire = "wire_12_";wire+=(fDetectorID+1000);}
     TString path = "/";path+=stat;path+="/";path+=plane;path+="/";path+=layer;path+="/";path+=wire;