+++ CHANGELOG.md
@@ -68,6 +68,7 @@ it in future.
 * build(field,nutaudet): remove unnecessary ROOT_INCLUDE_DIR include
 * feat(geometry): Remove 2018, 2022 global designs
 * run_simScript.py: remove --muShieldDesign flag
+* Remove unused straw veto station
 
 ## 25.01